Input: xpad - add Lenovo Legion Go controllers
When the controllers are both attached (or detached), they appear as
0x17ef, 0x6182.  When only one is attached, they appear as
0x17ef, 0x6184.

84 (mixed attachment) is already recognized as a gamepad by Linux.  This
patch adds 82 (both attached/detached).

These controllers have many more buttons than is typical.  With this
patch, here are the functioning buttons when it presents at 82.

Recognized:

- X, Y, A, B
- both analog sticks, including L3/R3
- D-pad
- menu and capture buttons

Not yet recognized:

- start, select
- 4 rear paddle buttons (Y1, Y2, Y3, M3)

(There are also non-functional buttons when reporting as 84 that are
out-of-scope for this patch.)